% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{byars_ci}
\alias{byars_ci}
\title{Byar's confidence interval}
\usage{
byars_ci(o, n, ci = 0.95)
}
\arguments{
\item{o}{observed or numerator value}

\item{n}{expected or denominator value}

\item{ci}{confidence interval coverage required.  Default is 0.95 for 95\\%
confidence interval}
}
\value{
A data.frame with Rate =  o/n, lower confidence interval limit, and
upper confidence interval limit
}
\description{
Byar's confidence interval for counts, crude rates or
indirectly standardised ratios
}
\examples{
byars_ci(50, 120)
}
