% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_sql_data_types.R
\name{derive_sql_data_types}
\alias{derive_sql_data_types}
\title{Derive SQL Data Types}
\usage{
derive_sql_data_types(df, buffer = 0)
}
\arguments{
\item{df}{R dataframe}

\item{buffer}{use to increase size of varchar fields - default is 0.}
}
\value{
Creates a vector of column names and derived SQL data types.
Can be used within data types parameter of dbWriteTable() from DBI package.
}
\description{
Derive equivalent SQL data types from an R dataframe
}
\examples{
derive_sql_data_types(mtcars)
}
