% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{f_year}
\alias{f_year}
\title{Financial year from date}
\usage{
f_year(d, short = FALSE)
}
\arguments{
\item{d}{a date, or vector of dates.}

\item{short}{TRUE/FALSE value, where TRUE returns 25/26 format, and FALSE
returns 2025/26. Default is FALSE.}
}
\value{
a character value, or vector, of the form  24/25 for fiscal
year 2024/25
}
\description{
Financial year from date
}
\examples{
f_year(Sys.Date())
f_year(Sys.Date(), short = TRUE)
}
