% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{f_year_end}
\alias{f_year_end}
\title{Financial year end date}
\usage{
f_year_end(d)
}
\arguments{
\item{d}{a date, or vector of dates.}
}
\value{
a date, or vector of dates for the 31st March in the relevant
financial year
}
\description{
Financial year end date
}
\examples{
f_year_end(Sys.Date())
}
