% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{f_year_start}
\alias{f_year_start}
\title{Financial year start date}
\usage{
f_year_start(d)
}
\arguments{
\item{d}{a date, or vector of dates.}
}
\value{
a date, or vector of dates for the 1st April in the relevant
financial year
}
\description{
Financial year start date
}
\examples{
f_year_start(Sys.Date())
}
