% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRs.R
\name{ISR_deprivation}
\alias{ISR_deprivation}
\title{Calculate Indirectly Standardised ratio for deprivation quintiles, standardised by age and/or sex}
\usage{
ISR_deprivation(.dt, age = TRUE, sex = FALSE)
}
\arguments{
\item{.dt}{A data.frame contain data you need to calculate the table with the following column names:
\itemize{
\item numerator - the count of events in question
\item denominator - the population or group at risk, or group/segment
\item imd_code - numeric encoded IMD quintile, with unknown coded as 999
\item age_group_code - age bands used for standardisation.  Bin/age band size is not important provided it is consistent.
\item sex - numeric encoded sex
}}

\item{age}{TRUE/FALSE whether to include age in the standardisation.}

\item{sex}{TRUE/FALSE whether to include sex in the standardisation.}
}
\value{
a data.frame with
}
\description{
This function uses negative binomial regression, to counter overdispersion, to adjust data by calculating and expected rate of events, which is compared to the observed event: observed / expected.
A ratio of 1 means observed = expected, <1 means observed < expected, and >1 means observed > expected.
}
\examples{
# This example won't work at the moment
\dontrun{
  ISR_deprivation(a)
  }

}
