% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISRs.R
\name{ISR_deprivation_plot}
\alias{ISR_deprivation_plot}
\title{Plot function for output parameters of ISR calculation}
\usage{
ISR_deprivation_plot(
  .dt,
  description = "Myocardial Infarction (Under 75yrs) - 2024/25",
  x_scale = seq(0, 1.5, 0.1)
)
}
\arguments{
\item{.dt}{A data.frame, generated by ISR_deprivation}

\item{description}{text to be included in the plot title, default example for MI admissions in 2024/25}

\item{y_scale}{vector of values, such as the default generated by \code{seq}.  Default around the 0 - 1.2 range}
}
\value{
A ggplot object (list)
}
\description{
Plot function for output parameters of ISR calculation
}
\examples{
\dontrun{
standardised_dep <- ISR(my_dt)
ISR_deprivation_plot()
}
}
