% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{theme_icb}
\alias{theme_icb}
\title{Automatic ggplot2 ICB theme}
\usage{
theme_icb(base_size = 12, base_family = "Open Sans")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
A theme designed to complement the ICB colour palette.
Uses a light, neutral background and minimalist styling.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
  ggplot2::geom_point(aes(colour = mpg)) +
  scale_colour_ICB_gradient() +
  theme_ICB()

}
