% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_binomial.R
\name{thin_binomial}
\alias{thin_binomial}
\title{Binomial thinning}
\usage{
thin_binomial(x, p)
}
\arguments{
\item{x}{vector of counts (e.g. arrivals per day)}

\item{p}{vector of probabilities (e.g. arrivals get admitted)}
}
\value{
vector of thinned counts
}
\description{
Binomial thinning
}
\examples{
set.seed(2024)
x_counts <- c(0, 3, 10, 5, 2)
p        <- 0.4
thin_binomial(x_counts, p)
}
