% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSOLutils-package.R
\docType{package}
\name{BSOLutils-package}
\alias{BSOLutils}
\alias{BSOLutils-package}
\title{BSOLutils: Utilities and general helper R package with various functions}
\description{
This package contains numerous helper functions. There include functions for confidence intervals, predicting SQL data types, ggplot themes and palettes, ploty helpers, binomial thinning, standardisation methods and dispersion metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://birmingham-and-solihull-ics.github.io/BSOLutils/}
  \item Report bugs at \url{https://github.com/birmingham-and-solihull-ics/BSOLutils/issues,}
}

}
\author{
\strong{Maintainer}: Chris Mainey \email{c.mainey1@nhs.net} (\href{https://orcid.org/0000-0002-3018-6171}{ORCID})

Other contributors:
\itemize{
  \item Siti Hassan \email{siti.hassan@nhs.net} [contributor]
  \item Archie Atack \email{a.atack@nhs.net} [contributor]
}

}
\keyword{internal}
