% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSOLutils-package.R
\docType{package}
\name{BSOLutils-package}
\alias{BSOLutils}
\alias{BSOLutils-package}
\title{BSOLutils: A utilities and general helper R package with various functions}
\description{
This package contains numerous helpfer functions. There include confidence intervals, binomial thinning, standardisation methods and dispersion metrics.
}
\author{
\strong{Maintainer}: Chris Mainey \email{c.mainey1@nhs.net} (\href{https://orcid.org/0000-0002-3018-6171}{ORCID})

}
\keyword{internal}
